+++ common/handlers/loader_test.go
@@ -6,36 +6,35 @@ import (
 	"testing"
 
 	log "github.com/sirupsen/logrus"
-
-	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/core/worker"
-	trace "gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/protos/trace"
+	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/engine"
+	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/protos/envelope"
 )
 
 type MockUnmarshaller struct {
 	t *testing.T
 }
 
-func (u *MockUnmarshaller) Unmarshal(msg interface{}, t *trace.Trace) error {
+func (u *MockUnmarshaller) Unmarshal(msg interface{}, envelope *envelope.Envelope) error {
 	if msg.(string) == "error" {
 		return fmt.Errorf("Could not unmarshall")
 	}
 	return nil
 }
 
-func makeLoaderContext(i int) *worker.Context {
-	ctx := worker.NewContext()
-	ctx.Reset()
-	ctx.Prepare([]worker.HandlerFunc{}, log.NewEntry(log.StandardLogger()), nil)
+func makeLoaderContext(i int) *engine.TxContext {
+	txctx := engine.NewTxContext()
+	txctx.Reset()
+	txctx.Prepare([]engine.HandlerFunc{}, log.NewEntry(log.StandardLogger()), nil)
 
 	switch i % 2 {
 	case 0:
-		ctx.Msg = "error"
-		ctx.Keys["errors"] = 1
+		txctx.Msg = "error"
+		txctx.Keys["errors"] = 1
 	case 1:
-		ctx.Msg = "valid"
-		ctx.Keys["errors"] = 0
+		txctx.Msg = "valid"
+		txctx.Keys["errors"] = 0
 	}
-	return ctx
+	return txctx
 }
 
 func TestLoader(t *testing.T) {
@@ -43,16 +42,16 @@ func TestLoader(t *testing.T) {
 	loader := Loader(&mu)
 
 	rounds := 10
-	outs := make(chan *worker.Context, rounds)
+	outs := make(chan *engine.TxContext, rounds)
 	wg := &sync.WaitGroup{}
 	for i := 0; i < rounds; i++ {
 		wg.Add(1)
-		ctx := makeLoaderContext(i)
-		go func(ctx *worker.Context) {
+		txctx := makeLoaderContext(i)
+		go func(txctx *engine.TxContext) {
 			defer wg.Done()
-			loader(ctx)
-			outs <- ctx
-		}(ctx)
+			loader(txctx)
+			outs <- txctx
+		}(txctx)
 	}
 	wg.Wait()
 	close(outs)
@@ -62,8 +61,8 @@ func TestLoader(t *testing.T) {
 
 	for out := range outs {
 		errCount := out.Keys["errors"].(int)
-		if len(out.T.Errors) != errCount {
-			t.Errorf("Loader: expected %v errors but got %v", errCount, out.T.Errors)
+		if len(out.Envelope.Errors) != errCount {
+			t.Errorf("Loader: expected %v errors but got %v", errCount, out.Envelope.Errors)
 		}
 	}
 }