+++ common/handlers/marker.go
@@ -2,19 +2,19 @@ package handlers
 
 import (
 	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/core/services"
-	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/core/worker"
+	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/engine"
 )
 
 // Marker creates an handler that mark offsets
-func Marker(offset services.OffsetMarker) worker.HandlerFunc {
-	return func(ctx *worker.Context) {
+func Marker(offset services.OffsetMarker) engine.HandlerFunc {
+	return func(txctx *engine.TxContext) {
 		// Marker is expected to be registered as one of the firt handlers so we are sure we alway mark messages
-		ctx.Next()
+		txctx.Next()
 
 		// Mark message
-		err := offset.Mark(ctx.Msg)
+		err := offset.Mark(txctx.Msg)
 		if err != nil {
-			ctx.Error(err)
+			txctx.Error(err)
 		}
 	}
 }