+++ common/handlers/marker_test.go
@@ -5,7 +5,7 @@ import (
 	"sync"
 	"testing"
 
-	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/core/worker"
+	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/engine"
 )
 
 type MockOffsetMarker struct {
@@ -19,18 +19,18 @@ func (o *MockOffsetMarker) Mark(msg interface{}) error {
 	return nil
 }
 
-func makeMarkerContext(i int) *worker.Context {
-	ctx := worker.NewContext()
-	ctx.Reset()
+func makeMarkerContext(i int) *engine.TxContext {
+	txctx := engine.NewTxContext()
+	txctx.Reset()
 	switch i % 2 {
 	case 0:
-		ctx.Msg = "error"
-		ctx.Keys["errors"] = 1
+		txctx.Msg = "error"
+		txctx.Keys["errors"] = 1
 	case 1:
-		ctx.Msg = "valid"
-		ctx.Keys["errors"] = 0
+		txctx.Msg = "valid"
+		txctx.Keys["errors"] = 0
 	}
-	return ctx
+	return txctx
 }
 
 func TestMarker(t *testing.T) {
@@ -38,16 +38,16 @@ func TestMarker(t *testing.T) {
 	marker := Marker(&mo)
 
 	rounds := 100
-	outs := make(chan *worker.Context, rounds)
+	outs := make(chan *engine.TxContext, rounds)
 	wg := &sync.WaitGroup{}
 	for i := 0; i < rounds; i++ {
 		wg.Add(1)
-		ctx := makeMarkerContext(i)
-		go func(ctx *worker.Context) {
+		txctx := makeMarkerContext(i)
+		go func(txctx *engine.TxContext) {
 			defer wg.Done()
-			marker(ctx)
-			outs <- ctx
-		}(ctx)
+			marker(txctx)
+			outs <- txctx
+		}(txctx)
 	}
 	wg.Wait()
 	close(outs)
@@ -58,8 +58,8 @@ func TestMarker(t *testing.T) {
 
 	for out := range outs {
 		errCount := out.Keys["errors"].(int)
-		if len(out.T.Errors) != errCount {
-			t.Errorf("Marker: expected %v errors but got %v", errCount, out.T.Errors)
+		if len(out.Envelope.Errors) != errCount {
+			t.Errorf("Marker: expected %v errors but got %v", errCount, out.Envelope.Errors)
 		}
 	}
 }