+++ common/handlers/producer.go
@@ -2,16 +2,16 @@ package handlers
 
 import (
 	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/core/services"
-	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/core/worker"
+	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/engine"
 )
 
 // Producer creates a producer handler
-func Producer(p services.Producer) worker.HandlerFunc {
-	return func(ctx *worker.Context) {
-		// Produce trace protobuffer
-		err := p.Produce(ctx.T)
+func Producer(p services.Producer) engine.HandlerFunc {
+	return func(txctx *engine.TxContext) {
+		// Produce Envelope
+		err := p.Produce(txctx.Envelope)
 		if err != nil {
-			ctx.Error(err)
+			txctx.Error(err)
 		}
 	}
 }