+++ common/infra/marshaller.go
@@ -4,21 +4,21 @@ import (
 	"fmt"
 
 	"github.com/gogo/protobuf/proto"
-	trace "gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/protos/trace"
+	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/protos/envelope"
 )
 
-// TracePbMarshaller assumes that message is a trace protobuf
-type TracePbMarshaller struct{}
+// EnvelopeMarshaller assumes that message is an Envelope
+type EnvelopeMarshaller struct{}
 
-// Marshal Trace into a message assumed to be a protobuf
-func (u *TracePbMarshaller) Marshal(t *trace.Trace, msg interface{}) error {
-	// Cast message into trace protobuf
-	pb, ok := msg.(*trace.Trace)
+// Marshal Envelope into a message assumed to be a protobuf
+func (u *EnvelopeMarshaller) Marshal(t *envelope.Envelope, msg interface{}) error {
+	// Cast message into Envelope
+	pb, ok := msg.(*envelope.Envelope)
 	if !ok {
 		return fmt.Errorf("Message does not match expected format")
 	}
 
-	// Merge msg into trace
+	// Merge msg into Envelope
 	proto.Merge(pb, t)
 
 	return nil