+++ protos/envelope/envelope.go
@@ -1,9 +1,9 @@
-package trace
+package envelope
 
 import (
 	common "gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/protos/common"
 )
 
-func (t *Trace) Error() string {
+func (t *Envelope) Error() string {
 	return common.Errors(t.Errors).Error()
 }